feat: export gnoland integration function
